requirements.txt
dgl==0.6.1
dgl-cu102==0.7.2
jarvis-tools==2021.12.16
joblib==1.1.0
numpy==1.21.1
pymatgen==2022.1.9
scipy==1.7.0
sympy==1.8
torch==1.10.1
tqdm==4.60.0
